+++ packages/localize/testdata/xliff-init/goldens/foo.ts
@@ -1,3 +1,3 @@
 import {msg} from '../../../lit-localize.js';
 
-msg('string', 'Hello World!');
+msg('Hello World!', {id: 'string'});